+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 ==========
 
+**February 11, 2015**
++ 1.0.1
+  + Added horizontal and vertical margin compensation
+  + Images scale according to font size, (mimicking ems)
+  + Organized and refactored some code
+
 **November 12, 2014**
 + 1.0.0
   + Actually a plugin